pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glasstape"
version = "1.0.0"
description = "Zero-trust runtime governance for AI agents"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "GlassTape Team", email = "meetharsharora@proton.me"}
]
maintainers = [
    {name = "GlassTape Team", email = "meetharsharora@proton.me"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "governance", "policy", "security", "agents", "llm"]
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
llm = [
    "openai>=1.0.0",
    "anthropic>=0.7.0",
]
# Note: CEL evaluation is built-in, no external dependencies needed

[project.urls]
Homepage = "https://github.com/glasstape/glasstape-sdk-python"
Documentation = "https://docs.glasstape.dev"
Repository = "https://github.com/glasstape/glasstape-sdk-python"
"Bug Tracker" = "https://github.com/glasstape/glasstape-sdk-python/issues"

[project.scripts]
glasstape = "glasstape.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["glasstape*"]

[tool.setuptools.package-data]
glasstape = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["glasstape"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["glasstape"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]